Enhance create_init function for wit initialization
Added a check to prevent reinitializing if .wit exists. Improved directory creation by removing 'exist_ok' parameter.

+++ wit/init.py
@@ -4,13 +4,19 @@ import os
 
 
 def create_init(path):
+    if os.path.exists('.wit'):
+        print("you already do init")
+        return
 
-    os.makedirs(f"{path}/.wit/staged_file", exist_ok = True)
-    os.makedirs(f"{path}/.wit/commits", exist_ok= True )
+    os.makedirs(f"{path}/.wit/staged_file")
+    os.makedirs(f"{path}/.wit/commits")
     with open(f"{path}/.wit/commits/head.txt","w") as file:
         file.write("")
 
     with open(f"{path}/.wit/.witignor.txt", "w") as file:
         file.write("")
+    print("the initial succses")
+
+
